fix some selections
the 'bracer' slot is not used, a search in itemtemplate for item type 30 will show that. Can also be confused with wrist slot (alot of wrist items named 'bracer')
Same with shield slot 31, no entires in DB with this item type. Instead i changed it to '11' which is left hand slot, and the item slot shields are used in. I think keeping shield in the list is handy though as 'left hand' might confuse someone as a choice when making a shield